requirements.txt
# AWS and S3 dependencies
boto3>=1.26.0
awswrangler>=3.0.0

# S3Tables
s3tables>=0.1.0

# AWS Glue dependencies (for local development)
pyspark>=3.3.0
pyglue>=0.4.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=12.0.0

# Configuration management
pyyaml>=6.0
python-dotenv>=1.0.0

# Logging
structlog>=23.1.0

# Testing
pytest>=7.3.1
pytest-cov>=4.1.0
moto>=4.1.0  # AWS mocking

# Development tools
black>=23.3.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.3.0